% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clt.R
\name{clt}
\alias{clt}
\title{Monte Carlo Simulations for an Inverse CDF}
\usage{
clt(N = NULL, invcdf = NULL, B = 1000, mu = 0)
}
\arguments{
\item{N}{The sample sizes (number of draws) you would like to test,
must be a vector of integers greater than 1.}

\item{invcdf}{A string representing the unevaluated inverse cdf. This should include
the variable n (the sample size), typically enclosed by \code{runif()}, like so \code{runif(n).}}

\item{B}{The number of simulations to run for each sample.}

\item{mu}{The actual known mean of the pdf.}
}
\description{
\code{clt()} runs Monte Carlo simulations for a given inverse cdf. 
Outputs the z-score of individual simulations for each sample size 
to a data frame.
}
\examples{

# The below will sample a uniform distribution 10000 times, drawing 2, 10,
# 40, and 500 for each simulation and output a data frame.

df <- clt(N = c(2,10,40,500), invcdf = "runif(n, -1, 1)", B = 10000)

}
