% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clt_anim.R
\name{clt_anim}
\alias{clt_anim}
\title{Making animated gifs of each N}
\usage{
clt_anim(data = NULL, ci = 1.96, xlim = c(-4, 4), compile = FALSE,
  fps = 4, filename = "clt.gif")
}
\arguments{
\item{ci}{The confidence interval you would like to evaluate at. Default is 1.96.}

\item{xlim}{A vector of the left and right x limits for the generated plots. For
certain distributions limits should be adjusted to more of the domain.}

\item{compile}{If TRUE, the function will compile into a gif then delete temp images,
if FALSE will output raw images to the working directory.}

\item{fps}{The number of frames per second for the gif.}

\item{filename}{The output name of the .gif file.}

\item{N}{The data frame to create a gif from.}
}
\description{
\code{clt_anim()} animates the output of \code{clt()}. This is an experimental function
and may not work in many cases. It creates ggplot histograms of the z-scores for 
each value of N, showing how N converges to a normal distribution as it increases.
}
\details{
NOTE: This function requires ImageMagick and uses the magick package.
}
