% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clt_eval.R
\name{clt_eval}
\alias{clt_eval}
\title{Evaluating Z-Scores}
\usage{
clt_eval(data = NULL, ci = 1.96, zonly = TRUE)
}
\arguments{
\item{data}{The data frame of z-scores to be evaluated.}

\item{ci}{The confidence interval you would like to evaluate at.}

\item{zonly}{If true, outputs only z-scores as a vector. If false,
outputs z-scores paired with N size as a data frame.}
}
\description{
\code{clt_eval()} evaluates the results from \code{clt()}. It outputs the 
percentage of z-scores which fell outside the specified critical values.
}
