% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_sample_means.R
\name{many_sample_means}
\alias{many_sample_means}
\title{Finds the means of a number of samples reps from the same numerical vector data}
\usage{
many_sample_means(vec, n, reps)
}
\arguments{
\item{vec}{A vector of numericals to sample from}

\item{n}{Sample size}

\item{reps}{Number of samples}
}
\value{
A vector of sample means
}
\description{
Finds the means of a number of samples reps from the same numerical vector data
}
