% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mean.R
\name{sample_mean}
\alias{sample_mean}
\title{Takes the mean of a sample of numerical vector data}
\usage{
sample_mean(vec, n)
}
\arguments{
\item{vec}{A vector of numericals to sample from}

\item{n}{Sample size}
}
\value{
The mean of a sample
}
\description{
Takes the mean of a sample of numerical vector data
}
