% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_means_ns.R
\name{sample_means_ns}
\alias{sample_means_ns}
\title{Finds any number of sample means reps across any number of sample sizes n}
\usage{
sample_means_ns(vec, reps, ns)
}
\arguments{
\item{vec}{A vector of numericals to sample from}

\item{reps}{Number of samples}

\item{ns}{A vector of different sample sizes}
}
\value{
A dataframe consisting of sample sizes and their respective sample means
}
\description{
Finds any number of sample means reps across any number of sample sizes n
}
